Refactor authentication and add logout functionality
The authentication system is overhauled for better usability and additional features. A new logout function is added to AuthService. UserClaimsPrincipalFactory has been updated to include user roles and given name. Removed unnecessary/dead lines of code in Login.tsx and AppClaimsPrincipalFactory.cs for clean up. Authentication checks are also now implemented in AuthController to redirect authenticated users to the dashboard. The Index.tsx page is updated to include options for authenticated users to navigate to the dashboard and for unauthenticated users to log in or register. The change is necessary to improve user experience and the security of the application.