Refactor HomeController and prepare for profile update
HomeController was refactored by removing the ILogger dependency as it was not necessary. Also, the [`UpdateProfileRequest`] class was added and used in the ProfileController. Files and code related to 'Auth' were moved into 'Requests' namespace for better organization. The next step would be implementing the functionality for updating user profile.