Implement toast notifications and refactor email service
Added 'toast' notifications to the UI to improve user experience. The toast component uses the 'react-hot-toast' package to display success, error, and info messages to users. Integrated toasts into 'Auth' and 'Guest' layouts to allow all authenticated and guest pages display messages through the use of the 'usePage' hook.

Refactored 'EmailSender.cs' in the Infrastructure layer to include a 'LogEmail' method to remove repeated code. Adjusted 'AuthController' to replace 'amp;' from the callback url in confirmation emails to fix a routing issue. Also, modified the confirmation email process to redirect the user to the Dashboard after confirming their email and added a success toast message in this case.

Added 'ToastType.cs' in the Domain project and 'Toast.cs' in the Application project to define a consistent model for toasts throughout the app. Updated registration process in AuthService to generate a confirmation token and send a confirmation email to the user. Updated package.json and package-lock.json to include 'react-hot-toast' package.

This commit aims to improve the UX with the addition of toast notifications and fixes a routing issue in the email confirmation process. Moreover, it paves the way for future implementation of another functionality related to toasts due to the added toast model.