Add email confirmation functionality on email change
We added the functionality that when a user changes their email, a confirmation mail is sent to the new email address for verification. To achieve this, we refactored the `GenerateEmailConfirmationToken` method in `IAuthService` to `GenerateAndSendEmailConfirmationToken` to include email sending feature. Changes were also made to `ProfileController` and `AuthService` to make use of this new method. Corresponding methods that trigger the delivery of confirmation email in `AuthController` and `ConfirmEmail` were refactored accordingly.