Refactor template classifications and sources in template.json file
The commit changes the classifications and tags in the template.json file. This was done to more accurately reflect the new structure and type of the project. The sources were also updated to exclude unnecessary files from the template, this change was implemented to improve project cleanliness and avoid clutter. Finally, conditions were added to dynamically exclude the sources based on the type of client chosen for the project. This was done to ensure that only the necessary files would be included depending on the chosen client (React, Vue, Headless), enhancing the project's modularity.